refactor: reorganize project structure
* Move webhook code from pkg/webhook to internal/webhook
* Move main application to cmd/webhook
* Reorganize test files:
  - Move test manifests to test/e2e
  - Move integration scripts to test/integration
* Update import paths and file references
* Update build configuration files for new structure

This change better follows Go project layout conventions:
- cmd/ for main applications
- internal/ for private application code
- pkg/ reserved for reusable packages
- test/ for organized test files

No functional changes were made.

run-integ-test
