+++ .gitignore
@@ -26,3 +26,6 @@ dist
 
 # AI
 flat/
+
+# repomix
+repomix-output.txt
