+++ .goreleaser.yml
@@ -10,7 +10,7 @@ builds:
     goarch:
       - amd64
       - arm64
-    main: ./pkg/webhook/cmd
+    main: ./cmd/webhook
     ldflags:
       - -s -w
       - -X main.version={{.Version}}
