+++ Dockerfile
@@ -3,7 +3,7 @@ FROM golang:1.23-alpine AS builder
 WORKDIR /app
 COPY . .
 RUN go mod download
-RUN CGO_ENABLED=0 GOOS=linux go build -o webhook ./pkg/webhook/cmd
+RUN CGO_ENABLED=0 GOOS=linux go build -o webhook ./cmd/webhook
 
 FROM alpine:3.21
 WORKDIR /
