+++ test/e2e/manifests/kind-config.yaml
