+++ test/integration/delete-kind-cluster.sh
