+++ test/integration/kind-deploy.sh
@@ -11,7 +11,7 @@ VERSION=${VERSION:-latest}
 kind export kubeconfig --name ${CLUSTER_NAME}
 kind load docker-image ${IMAGE_NAME}:${VERSION} --name ${CLUSTER_NAME}
 
-kubectl apply -f tests/manifests/webhook.yaml
+kubectl apply -f test/e2e/manifests/webhook.yaml
 kubectl wait --for=condition=Ready --timeout=60s -n ${NAMESPACE} certificate/pod-label-webhook-cert
-kubectl apply -f tests/manifests/deployment.yaml
+kubectl apply -f test/e2e/manifests/deployment.yaml
 kubectl wait --for=condition=Available --timeout=60s -n ${NAMESPACE} deployment/pod-label-webhook
